Feat: 헬스장별 리뷰 조회

+++ back/src/main/java/main012/server/gym/entity/GymReview.java
@@ -20,7 +20,7 @@ public class GymReview extends Auditable {
     @Column(name = "gym_review_id")
     private Long id;
 
-    private long gymGrade;
+    private int gymGrade;
 
     private String gymComment;
